feat: Prepare for the p2p handshake

+++ src/util/time.h
@@ -17,5 +17,5 @@ public:
 FORCE_INLINE uint64_t Time::Nanoseconds() noexcept {
     timespec now;
     clock_gettime(CLOCK_MONOTONIC, &now);
-    return uint64_t((now.tv_sec * 1'000'000'000L) + now.tv_nsec);
+    return (uint64_t)now.tv_sec * 1'000'000'000ULL + (uint64_t)now.tv_nsec;
 }
